+++ CHANGELOG.rst
@@ -18,9 +18,9 @@ Each individual change should have a link to the pull request after the descript
 
 1.1.2 (2024-02-07)
 ------------------
-
 Added
 ^^^^^
+- Update OneHotEncodingTransformer to default to returning int8 columns `#175 <https://github.com/lvgig/tubular/pull/175>`_
 - Updated NullIndicator to return int8 columns `#173 https://github.com/lvgig/tubular/pull/173`_
 
 1.1.1 (2024-01-18)